# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    LanguageLoc:
      type: object
      properties:
        language:
          type: string
        mainLines:
          type: integer
          format: int64
        testLines:
          type: integer
          format: int64
    LocEntry:
      type: object
      properties:
        repositoryId:
          type: string
        commitSha:
          type: string
        status:
          $ref: "#/components/schemas/LocStatus"
        languages:
          type: array
          items:
            $ref: "#/components/schemas/LanguageLoc"
    LocListResponse:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/LocEntry"
...
    LocStatus:
      type: string
      enum:
      - COUNTED
      - PENDING
      - EMPTY
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-platform-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
  /githost/api/loc:
    get:
      summary: Lines of code of every repository's default branch
      description: "One row per repository this host holds, or per named repositoryId\
        \ it holds. COUNTED rows carry the numbers; PENDING rows are queued for counting\
        \ by this call; EMPTY rows have no commit yet."
      operationId: listLoc
      parameters:
      - name: repositoryId
        in: query
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocListResponse"
      tags:
      - Loc List Resource
...
info:
  title: qits-githost API
  version: 2026.1002.121335